% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dates_from_tiff_files.R,
%   R/get.Series.mean.R, R/get.Series.median.R
\name{extract_dates_from_tiff_files}
\alias{extract_dates_from_tiff_files}
\title{extract_dates_from_tiff_files}
\usage{
extract_dates_from_tiff_files(tiff_files)

extract_dates_from_tiff_files(tiff_files)

extract_dates_from_tiff_files(tiff_files)
}
\arguments{
\item{tiff_files}{List of files in .tif format}
}
\value{
The names of each of the files in the formats \%Y-\%m-\%d or YYYYYYYMMDD

The names of each of the files in the formats \%Y-\%m-\%d or YYYYYYYMMDD

The names of each of the files in the formats \%Y-\%m-\%d or YYYYYYYMMDD
}
\description{
extract_dates_from_tiff_files

extract_dates_from_tiff_files

extract_dates_from_tiff_files
}
