% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM.plot.R
\name{GAM.plot}
\alias{GAM.plot}
\title{GAM.plot}
\usage{
GAM.plot(
  data,
  doy_col = "DOY",
  year_col = "Year",
  proporcion_col = "Proporcion_Missing",
  prediccion_col = "Predicted",
  sos = NULL,
  eos = NULL,
  max_doy = NULL
)
}
\arguments{
\item{data}{A DataFrame that contains the data to be plotted.}

\item{doy_col}{Name of the column representing the Julian Day (DOY).}

\item{year_col}{Name of the column representing the Year.}

\item{proporcion_col}{Name of the column representing the Proportion of Missing Data.}

\item{prediccion_col}{Name of the column representing the Prediction.}

\item{sos}{Optional value for the SOS line.}

\item{eos}{Optional value for the EOS line.}

\item{max_doy}{Optional value for the MAX line.}
}
\value{
Object of type ggplot that shows the results of the prediction obtained with the GAM.missing function.
}
\description{
GAM.plot
}
\examples{
proportions <- data.frame(
  Year = rep(2020L, 5), DOY = c(20, 50, 80, 110, 140),
  Proporcion_Missing = c(0.1, 0.2, 0.3, 0.2, 0.1),
  Predicted = c(0.12, 0.18, 0.25, 0.18, 0.12)
)
p <- suppressWarnings(GAM.plot(proportions))
inherits(p, "ggplot")
}
