% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.Quality.R
\name{general.Quality}
\alias{general.Quality}
\title{general.Quality}
\usage{
general.Quality(df, period_col, observations_col)
}
\arguments{
\item{df}{Data base generated from the use of the function quality.Series}

\item{period_col}{Column where the periods are stored (every 91 days)}

\item{observations_col}{Columna donde se almacenan el numero de observaciones por periodo y}
}
\value{
A list of files that stores a new data frame and an interactive graphic.
}
\description{
general.Quality
}
