% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrplot.R
\name{mrplot}
\alias{mrplot}
\title{mrplot}
\usage{
mrplot(
  data,
  doy_col = "DOY",
  serie_col = "NDVI",
  colors = c("green4", "orange")
)
}
\arguments{
\item{data}{Dataframe type object to store the time series}

\item{doy_col}{Column with Julian days}

\item{serie_col}{Column that stores the variable of interest}

\item{colors}{Selection of the colors of the graphic}
}
\value{
Graph showing the relationship between "NA" and "NO NA" observations with Julian days
}
\description{
mrplot
}
\examples{

mrplot(Zaidin,doy_col = "DOY",serie_col = "NDVI_median", colors = c ("green4", "orange4"))

}
