% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renames.image.VPP.R
\name{renames.image.VPP}
\alias{renames.image.VPP}
\title{renames.image.VPP}
\usage{
renames.image.VPP(...)
}
\arguments{
\item{...}{Legacy arguments.}
}
\value{
Does not return a value, but renames the TIF files in the specified directory.
}
\description{
Rename TIF files in a directory according to the year extracted from the name.
}
\details{
This function searches for TIF files in a specified directory and
renames them using the year extracted from the filename.
The new file name shall be in the format 'yyyy.tif'.
}
