% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.HRVPP.R
\name{Download.HRVPP}
\alias{Download.HRVPP}
\title{Download.HRVPP}
\arguments{
\item{user}{User in which it is used in the WEkEO platform}

\item{password}{Password used by the user on the WEkEO platform}

\item{dataset_id}{Database where the product to be downloaded is stored, see WEkEO documentation. En el cso de los productos HRVPP usar: "EO:EEA:DAT:CLMS_HRVPP_VPP"}

\item{productType}{Select the type of product to download (example: Start of Season_date(SOSD)). For other product types see WEkEO documentation.}

\item{tileId}{ndicate the tesserae (mosaic) you want to download}

\item{productGroupID}{The product group will be selected, which can be Season 1 or 2 (example: s2).}

\item{start}{Start date with structure: "yyyy-mm-ddT00:00:00:00.000Z".}

\item{end}{End date with structure: "yyyy-mm-ddT00:00:00:00.000Z".}

\item{bbox}{Coordinates of the study area: Xmin,Ymin,Xmax,Ymax}

\item{download_path}{Directory where the .tif files will be stored}
}
\value{
Set of .tif files relating to the selected HRVPP product(s)
}
\description{
Download.HRVPP
}
\examples{
#' Download.HRVPP(
user = "user_name",
password = "Password",
dataset_id = "EO:EEA:DAT:CLMS_HRVPP_VPP",
productType = "SOSD",
productGroupId = "s1",
tileId = "30TXL",
start = "2020-01-01T00:00:00.000Z",
end = "2020-01-10T00:00:00.000Z",
bbox = c(-0.89285, 41.48762, -0.86284, 41.50456),
download_path = "local_directory"
)
}
