% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM.missing.R
\name{GAM.missing}
\alias{GAM.missing}
\title{GAM.missing}
\usage{
GAM.missing(data, year_col, doy_col, missing_col)
}
\arguments{
\item{data}{Data frame containing the time series information of the vegetation index with presence of missing data}

\item{year_col}{Column storing the information of the years corresponding to each time series (yyyy)}

\item{doy_col}{Column that stores the information related to the Julian days}

\item{missing_col}{}
}
\value{
A list object where the prediction results and the GAM model data are stored.
}
\description{
GAM.missing
}
\examples{
resultado <- GAM.missing(data, "Year", "DOY", "NDVI_median")
print(resultado$proporciones)
summary(resultado$model)

}
