% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Series.mean.R
\name{get.Series.mean}
\alias{get.Series.mean}
\title{get.Series.mean}
\usage{
get.Series.mean(pathRaster = NULL, shapefile = NULL, factorR = NULL)
}
\arguments{
\item{pathRaster}{Path to the directory containing raster files in TIFF format.}

\item{shapefile}{sf' object defining the areas of interest for data mining.}

\item{factorR}{Factor scale for reflectance bands. It is a numerical value that is used
to adjust the extracted values.}
}
\value{
A data frame containing the ID of the areas of interest, the date corresponding to each
#' TIFF file and the adjusted mean reflectance values.
}
\description{
Extracts and processes reflectance indices from TIFF raster files using a shapefile.
}
\details{
This function allows the extraction of mean band reflectance values from raster images in TIFF format.
using a shapefile to define the areas of interest. The extracted values are processed and
returned as a data frame.
}
\examples{
# Extracting indexes from a raster directory and a shapefile
índices <- extraer_índices("ruta/a/carpeta", shapefile = mi_shapefile, factorR = 10000)
}
