% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Series.median.R
\name{get.Series.median}
\alias{get.Series.median}
\title{get.Series.median}
\arguments{
\item{pathRaster}{Path to the directory containing the raster files in TIFF format.}

\item{shapefile}{An object of class 'sf' defining the area of interest for extraction.}

\item{factorR}{A number representing the scale factor for the reflectance bands.}
}
\value{
A data frame containing three columns: ID (shapefile identifier),
Date (date corresponding to the raster file) and Median (median of the extracted values).
}
\description{
Get Series from Raster Files
}
\details{
This function extracts time series reflectance data from raster files in TIFF format
and calculates the median of the values within an area defined by a shapefile.
}
\examples{
Example of use: #' #' result <- get_series(pathRaster = "path/to/them/tifs",
#' shapefile = my_shapefile, #' factorR = 10000)
}
