% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Stack.R
\name{get.Stack}
\alias{get.Stack}
\title{get.Stack}
\usage{
get.Stack(IV_path, QFLAG, output_path)
}
\arguments{
\item{IV_path}{Path to the directory containing VI files in .tif format.}

\item{QFLAG}{Path to the directory containing QFLAG files.}

\item{output_path}{Path to the directory where the processed files will be saved.}
}
\value{
It does not return a value, but saves .tif files in the output directory
with the NDVI and QFLAG layers combined.
}
\description{
Processes vegetation index files and adds QFLAG layers.
}
\details{
This function searches for .TIFF files of vegetation indexes (VI) in a specified directory,
and for each file, it tries to find a corresponding QFLAG file.
If found, the QFLAG file is re-scaled to match
the resolution of the VI file and is added as a new layer.
Finally, the resulting file is saved in an output directory.
}
\examples{
# Processing VI files and adding QFLAG layers
get.Stack("path/to/VI", "path/to/QFLAG", "path/to/output")
}
