% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.Series.R
\name{quality.Series}
\alias{quality.Series}
\title{quality.Series}
\usage{
quality.Series(df, id_col, fecha_col, value_col)
}
\arguments{
\item{df}{Data base}

\item{id_col}{Column where the observation grid identifiers are located.}

\item{fecha_col}{Column where the date is located (\%yyyyy\%mm\%dd)}

\item{value_col}{Column from which the quality of the time series will be determined (estimating the number of observations).}
}
\value{
A list of files where a dataframe with the results of the quality assessment and an interactive graphic are stored.
}
\description{
quality.Series
}
\examples{
Estimar la calidad  series temporales de NDVI en diferentes parcelas
quality.Series(df = data, id_col = "ID", fecha_col = "Date", value_col = "NDVI")
}
