% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renames.image.IV.R
\name{renames.image.IV}
\alias{renames.image.IV}
\title{renames.image.IV}
\usage{
renames.image.IV(input_folder = NULL)
}
\arguments{
\item{input_folder}{Path to the directory containing TIF files.}
}
\value{
It does not return a value, but renames the TIF files in the specified directory.
}
\description{
Rename TIF files in a directory according to the date contained in the name.
}
\details{
This function looks for TIF files in a specified directory and
rename them using the date extracted from the file name.
The new file name will have the format 'yyyymmdd.tif'.
}
\examples{
# Rename TIF files in the specified folder
rename_files("path/to/folder")
}
