% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSapp.R
\name{runTSapp}
\alias{runTSapp}
\title{Run Shiny applications from the TSGenerator package}
\usage{
runTSapp(app = "DownloadVI")
}
\arguments{
\item{app}{Character string. Specifies which Shiny application to run.
Valid options: "DownloadVI", "DownloadSTPPI", "DownloadVPP", "getCleanVI",
"getStack", "renamesVI", "renamesVPP", "GAM". Por defecto es "DownloadVI".}
}
\value{
It does not return a value, but runs a Shiny application in the default web browser.
}
\description{
This function launches different Shiny applications that are part of the TSGenerator package.
Each application provides a graphical interface for performing different operations
related to time series.
}
\details{
The available applications are:
\itemize{
\item \code{DownloadVI}: Interface for downloading VI data (NDVI, fAPAR, LAI and PPI).
\item \code{DownloadSTPPI}: Interface for downloading PPI seasonal trajectory data (ST).
\item \code{DownloadVPP}: Interface for downloading VPP data.
\item \code{getCleanVI}: Data cleaning and processing tool.
\item \code{getStack}: Tool for stacking data sets.
\item \code{renamesVI}: VI file renaming utility.
\item \code{renamesVPP}: Utility to rename VPP files.
\item \code{GAM}: GAM model to analysis NA data
}
}
\examples{
\dontrun{
# Launch the DownloadVI application (default)
TSGenerator::runTSapp()

# Explicitly specify the DownloadVI application
TSGenerator::runTSapp("DownloadVI")

# Launch the getClean application
TSGenerator::runTSapp("getCleanVI")
}

}
