% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsImpute.R
\name{TsImpute}
\alias{TsImpute}
\title{TsImpute}
\usage{
TsImpute(data, group_col, value_col)
}
\arguments{
\item{data}{data frame where VI time series containing missing data are stored.}

\item{group_col}{column by which the grouping is to be performed (example: ID column)}

\item{value_col}{column where the time series of the VI data is stored (example: NDVI)}
}
\value{
A new file similar to the original data frame but with a new column where the imputed data is stored.
}
\description{
TsImpute
}
\examples{
#Example of application of the TsImpute function
new_Data <- TsImpute(df, group_col = "ID", value_col = "NDVI")
}
